chore: fold network policy types into environments
